Update Travis testing matrix (#679)
* Set up tox-travis

This will allow us to keep our matrix configuration in Tox instead of having to replicate it within Travis. Right now this does not fully match the Tox configuration, since it didn't previously, but that will be corrected later.

* Add testing on Python 3.5

This is in the supported versions and the Tox matrix, but not in the Travis matrix.

* Fix running flake8 and docs tox configurations

* Test on Python 3.7

* Add Python 3.7 to tox matrix

Forgot to do this, so Travis will know what environments to execute. The docs and flake8 remain at Python 3.6 since that's where they were before.